enlevage de cpusage, rame et nvidia + documentation : finie a 90% + nettoyage de l'API => si ca compile c'est que j'ai eu du bol ^_^

+++ desklet-rendering/src/rendering-desklet-mediaplayer.h
@@ -16,13 +16,13 @@ typedef struct {
 	
 	gdouble fDeskletWidth;
 	
-	gdouble fArtistWidth;
-	gdouble fArtistHeight;
+	gint fArtistWidth;
+	gint fArtistHeight;
 	gdouble fArtistXOffset;
 	gdouble fArtistYOffset;
 	
-	gdouble fTitleWidth;
-	gdouble fTitleHeight;
+	gint fTitleWidth;
+	gint fTitleHeight;
 	gdouble fTitleXOffset;
 	gdouble fTitleYOffset;
